requirements.txt
-e git+https://github.com/s1s5/fabmisc.git@c3229d4e5e942c26c792f45672597ad353aeb597#egg=fabmisc
Fabric==1.13.1
Jinja2==2.8.1
MarkupSafe==1.0
PyYAML==3.12
ansible==2.2.1.0
appdirs==1.4.3
asn1crypto==0.22.0
backports.ssl-match-hostname==3.5.0.1
cached-property==1.3.0
cffi==1.9.1
colorama==0.3.7
cryptography==1.8.1
docker-compose==1.11.2
docker-pycreds==0.2.1
docker==2.1.0
dockerpty==0.4.1
docopt==0.6.2
enum34==1.1.6
fabtools==0.20.0
functools32==3.2.3.post2
idna==2.5
ipaddress==1.0.18
jsonschema==2.6.0
packaging==16.8
paramiko==2.1.2
pyasn1==0.2.3
pycparser==2.17
pycrypto==2.6.1
pyparsing==2.2.0
requests==2.13.0
six==1.10.0
texttable==0.8.7
websocket-client==0.40.0
